+++ tests/Hover.test.cpp
@@ -1,0 +1,40 @@
+#include "doctest.h"
+#include "Fixture.h"
+
+TEST_SUITE_BEGIN("Hover");
+
+TEST_CASE_FIXTURE(Fixture, "show_string_length_on_hover")
+{
+    auto source = R"(
+        local x = "this is a string"
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "string (16 bytes)"));
+}
+
+TEST_CASE_FIXTURE(Fixture, "show_string_utf8_characters_on_hover")
+{
+    auto source = R"(
+        local x = "this is an emoji: 😁"
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "string (22 bytes, 19 characters)"));
+}
+
+TEST_SUITE_END();