+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Autocompletion items for items marked as `@deprecated` via documentation comments will now reflect their deprecated status
+- Show string literal byte length and utf8 characters on hover
 
 ### Changed